+++ app.R
@@ -23,7 +23,6 @@ source("source/tabs/bricksQ_tab.R")
 source("source/tabs/loomweight_tab.R")
 source("source/tabs/allfinds_tab.R")
 #source("source/tabs/sculpture_tab.R")
-#source("source/tabs/buildings_tab.R")
 
 # body
 source("source/body.R")
@@ -34,10 +33,10 @@ server <- function(input, output, session) {
 
   # show login dialog box when initiated
   showModal(login_dialog, session = getDefaultReactiveDomain())
-  # server code to handle basic setting, i.e. ip and pwd
+  # server code to handle the connection to field
   source('source/server/connect_serv.R', local = TRUE)
 
-  # server code to handle basic setting, i.e. ip and pwd
+  # server code to handle basic settings, i.e. project, trench/operation
   source('source/server/settings_serv.R', local = TRUE)
   # server code to import database, places etc.
   source('source/server/database_serv.R', local = TRUE)
@@ -53,7 +52,7 @@ server <- function(input, output, session) {
   # server code only for pottery quantification B form
   source('source/server/potteryQB_serv.R', local = TRUE)
 
-  # server code only for loomweights
+  # server code only for bricks
   source('source/server/bricks_serv.R', local = TRUE)
   source('source/server/bricksQ_serv.R', local = TRUE)
 
@@ -63,7 +62,6 @@ server <- function(input, output, session) {
 
 
   #source('source/server/sculpture_serv.R', local = TRUE)
-  #source('source/server/buildings_serv.R', local = TRUE)
 
 
   # uncomment this for distribution