+++ source/tabs/connect_tab.R
@@ -10,18 +10,17 @@ connect_tab <- tabItem(
                                      class = "login-text",
                                      textOutput("tab_connect.welcome_text",
                                                 container = tags$h1))))),
-      infoBox(width = 3, title = "Version", subtitle = "date: 22.10.2022",
-              icon = icon("code-branch"), value = "v.0.1.1", color = "olive")
+      infoBox(width = 3, title = "Version", subtitle = "date: 18.02.2023",
+              icon = icon("code-branch"), value = "v.0.2.2", color = "olive")
     ),
     fluidRow(
       box(width = 6,
           p("With this App, you can view and download various plots of data from
-      an iDAI.field-Database, that is otherwise usually inaccessible.
-      In order for the App to work, you need to have iDAI.field 2 or
-      Field Desktop running on your computer. Choose a project from
-      your Field Client from the selection below."),
-          p("The App is meant to be used with the milet-configuration and many
-            plots may only work for this configuration.")),
+      an iDAI.field/Field Desktop-Database. In order for the App to work,
+      you need to have iDAI.field 2 or Field Desktop running on your computer.
+      Choose a project from your Field Client from the selection below."),
+          p("The app is meant to be used with the milet-configuration and most
+            plots will only work with this configuration.")),
       box(title = "Please note", status = "warning",
           p("Large projects may take a while to load. Be prepared to wait
             after clicking 'Load Database'."))
@@ -44,11 +43,10 @@ connect_tab <- tabItem(
 )
 
 # define the ui of the login dialog box
-# will used later in server part
 login_dialog <- modalDialog(
   title = "Enter connection details to continue",
   footer = actionButton("tab_connect.connect","Connect"),
-  textInput("tab_connect.host","Host (iDAI.field Client)", value = "127.0.0.1"),
+  textInput("tab_connect.host","Host (Field Desktop)", value = "127.0.0.1"),
   textInput("tab_connect.user","Username", value = "username"),
   passwordInput("tab_connect.pwd","Password", value = "hallo"),
   tags$div(class = "warn-text",textOutput("tab_connect.error_msg"))