+++ NEWS.md
@@ -1,3 +1,11 @@
+# milQuant 0.2.2 _2023-02-18_
+
+* use idaifieldR v0.2.2 (fixing a problem where with an if-condition)
+* rudimentary form for bricks added
+* fix downloading plot in Finds, Loomweights and Bricks
+* fix error handling in connection (now actually displays the error)
+* WIP...
+
 # milQuant 0.2.1 _2022-05-15_
 
 * uses idaifieldR v0.2.1